+++ bridges/otelslog/handler.go
@@ -50,6 +50,7 @@ import (
 	"log/slog"
 	"runtime"
 	"slices"
+	"strings"
 
 	"go.opentelemetry.io/otel/log"
 	"go.opentelemetry.io/otel/log/global"
@@ -192,9 +193,11 @@ func (h *Handler) convertRecord(r slog.Record) log.Record {
 	if h.source {
 		fs := runtime.CallersFrames([]uintptr{r.PC})
 		f, _ := fs.Next()
+		funcName, namespace := splitFuncName(f.Function)
 		record.AddAttributes(
 			log.String(string(semconv.CodeFilepathKey), f.File),
-			log.String(string(semconv.CodeFunctionKey), f.Function),
+			log.String(string(semconv.CodeFunctionKey), funcName),
+			log.String(string(semconv.CodeNamespaceKey), namespace),
 			log.Int(string(semconv.CodeLineNumberKey), f.Line),
 		)
 	}
@@ -228,9 +231,8 @@ func (h *Handler) convertRecord(r slog.Record) log.Record {
 // Enable returns true if the Handler is enabled to log for the provided
 // context and Level. Otherwise, false is returned if it is not enabled.
 func (h *Handler) Enabled(ctx context.Context, l slog.Level) bool {
-	var param log.EnabledParameters
 	const sevOffset = slog.Level(log.SeverityDebug) - slog.LevelDebug
-	param.SetSeverity(log.Severity(l + sevOffset))
+	param := log.EnabledParameters{Severity: log.Severity(l + sevOffset)}
 	return h.logger.Enabled(ctx, param)
 }
 
@@ -475,4 +477,16 @@ func convert(v slog.Value) log.Value {
 		// have a "unhandled: " prefix than say that their code is panicking.
 		return log.StringValue(fmt.Sprintf("unhandled: (%s) %+v", v.Kind(), v.Any()))
 	}
+}
+
+// splitFuncName splits package path-qualified function name into
+// function name and package full name (namespace). E.g. it splits
+// "github.com/my/repo/pkg.foo" into
+// "foo" and "github.com/my/repo/pkg".
+func splitFuncName(f string) (string, string) {
+	i := strings.LastIndexByte(f, '.')
+	if i < 0 {
+		return "", ""
+	}
+	return f[i+1:], f[:i]
 }