+++ bridges/otelzap/core.go
@@ -36,6 +36,7 @@ package otelzap // import "go.opentelemetry.io/contrib/bridges/otelzap"
 import (
 	"context"
 	"slices"
+	"strings"
 
 	"go.uber.org/zap/zapcore"
 
@@ -142,8 +143,7 @@ func NewCore(name string, opts ...Option) *Core {
 
 // Enabled decides whether a given logging level is enabled when logging a message.
 func (o *Core) Enabled(level zapcore.Level) bool {
-	param := log.EnabledParameters{}
-	param.SetSeverity(convertLevel(level))
+	param := log.EnabledParameters{Severity: convertLevel(level)}
 	return o.logger.Enabled(context.Background(), param)
 }
 
@@ -178,8 +178,7 @@ func (o *Core) Sync() error {
 // Check determines whether the supplied Entry should be logged.
 // If the entry should be logged, the Core adds itself to the CheckedEntry and returns the result.
 func (o *Core) Check(ent zapcore.Entry, ce *zapcore.CheckedEntry) *zapcore.CheckedEntry {
-	param := log.EnabledParameters{}
-	param.SetSeverity(convertLevel(ent.Level))
+	param := log.EnabledParameters{Severity: convertLevel(ent.Level)}
 
 	logger := o.logger
 	if ent.LoggerName != "" {
@@ -202,10 +201,12 @@ func (o *Core) Write(ent zapcore.Entry, fields []zapcore.Field) error {
 
 	r.AddAttributes(o.attr...)
 	if ent.Caller.Defined {
+		funcName, namespace := splitFuncName(ent.Caller.Function)
 		r.AddAttributes(
 			log.String(string(semconv.CodeFilepathKey), ent.Caller.File),
 			log.Int(string(semconv.CodeLineNumberKey), ent.Caller.Line),
-			log.String(string(semconv.CodeFunctionKey), ent.Caller.Function),
+			log.String(string(semconv.CodeFunctionKey), funcName),
+			log.String(string(semconv.CodeNamespaceKey), namespace),
 		)
 	}
 	if ent.Stack != "" {
@@ -261,4 +262,16 @@ func convertLevel(level zapcore.Level) log.Severity {
 	default:
 		return log.SeverityUndefined
 	}
+}
+
+// splitFuncName splits package path-qualified function name into
+// function name and package full name (namespace). E.g. it splits
+// "github.com/my/repo/pkg.foo" into
+// "foo" and "github.com/my/repo/pkg".
+func splitFuncName(f string) (string, string) {
+	i := strings.LastIndexByte(f, '.')
+	if i < 0 {
+		return "", ""
+	}
+	return f[i+1:], f[:i]
 }